refactor: use partial views to handle rich text rendering
