--- src/Dfe.ContentSupport.Web/Services/Renderers/IContentRenderer.cs
@@ -1,7 +1,0 @@
-using Dfe.ContentSupport.Web.Models;
-
-namespace Dfe.ContentSupport.Web.Services;
-public interface IContentRenderer
-{
-    string Render(ContentSupportPage page);
-}
