--- src/Dfe.ContentSupport.Web/Services/Renderers/PartRenderers/OrderedListRenderer.cs
@@ -1,15 +1,0 @@
-using System.Text;
-
-namespace Dfe.ContentSupport.Web.Services;
-
-public class OrderedListRenderer : NodeRenderer
-{
-    public override void Render(dynamic node, StringBuilder sb)
-    {
-        sb.Append("<ol>");
-
-        ExploreNode(node, sb);
-
-        sb.Append("</ol>");
-    }
-}
