--- src/Dfe.ContentSupport.Web/Services/Renderers/PartRenderers/ParagraphRenderer.cs
@@ -1,15 +1,0 @@
-using System.Text;
-
-namespace Dfe.ContentSupport.Web.Services;
-
-public class ParagraphRenderer : NodeRenderer
-{
-    public override void Render(dynamic node, StringBuilder sb)
-    {
-        sb.Append("<p>");
-
-        ExploreNode(node, sb);
-
-        sb.Append("</p>");
-    }
-}
