--- src/Dfe.ContentSupport.Web/Services/Renderers/PartRenderers/TableCellRenderer.cs
@@ -1,13 +1,0 @@
-using System.Text;
-
-namespace Dfe.ContentSupport.Web.Services;
-
-public class TableCellRenderer : NodeRenderer
-{
-    public override void Render(dynamic node, StringBuilder sb)
-    {
-        sb.Append("<td>");
-        ExploreNode(node, sb);
-        sb.Append("</td>");
-    }
-}
