--- src/Dfe.ContentSupport.Web/Services/Renderers/PartRenderers/TableHeaderCellRenderer.cs
@@ -1,13 +1,0 @@
-using System.Text;
-
-namespace Dfe.ContentSupport.Web.Services;
-
-public class TableHeaderCellRenderer : NodeRenderer
-{
-    public override void Render(dynamic node, StringBuilder sb)
-    {
-        sb.Append("<th>");
-        ExploreNode(node, sb);
-        sb.Append("</th>");
-    }
-}
