--- src/Dfe.ContentSupport.Web/Services/Renderers/PartRenderers/TableRowRenderer.cs
@@ -1,13 +1,0 @@
-using System.Text;
-
-namespace Dfe.ContentSupport.Web.Services;
-
-public class TableRowRenderer : NodeRenderer
-{
-    public override void Render(dynamic node, StringBuilder sb)
-    {
-        sb.Append("<tr>");
-        ExploreNode(node, sb);
-        sb.Append("</tr>");
-    }
-}
