--- src/Dfe.ContentSupport.Web/Services/Renderers/PartRenderers/UnorderedListRenderer.cs
@@ -1,15 +1,0 @@
-using System.Text;
-
-namespace Dfe.ContentSupport.Web.Services;
-
-public class UnorderedListRenderer : NodeRenderer
-{
-    public override void Render(dynamic node, StringBuilder sb)
-    {
-        sb.Append("<ul>");
-
-        ExploreNode(node, sb);
-
-        sb.Append("</ul>");
-    }
-}
