+++ src/Dfe.ContentSupport.Web/Views/Shared/RichText/_HyperLink.cshtml
@@ -1,0 +1,12 @@
+@using Newtonsoft.Json.Linq
+@model JObject
+
+<a href="@Model["data"]["uri"]?.ToString()">
+    @if (Model["content"] is JArray childNodes)
+    {
+        foreach (var childNode in childNodes)
+        {
+            <partial name="RichText/_RichText" model="@childNode" />
+        }
+    }
+</a>
