+++ src/Dfe.ContentSupport.Web/Views/Shared/RichText/_RichText.cshtml
@@ -1,0 +1,46 @@
+@using Dfe.ContentSupport.Web.Common
+@using Newtonsoft.Json.Linq
+@model JObject
+
+@{
+    var nodeType = Model["nodeType"]?.ToString();
+
+    switch (nodeType)
+    {
+        case RichTextTags.Document:
+            <partial name="RichText/_Document" model="@Model" />
+            break;
+        case RichTextTags.Paragraph:
+            <partial name="RichText/_Paragraph" model="@Model" />
+            break;
+        case RichTextTags.UnorderedList:
+            <partial name="RichText/_UnorderedList" model="@Model" />
+            break;
+        case RichTextTags.OrderedList:
+            <partial name="RichText/_OrderedList" model="@Model" />
+            break;
+        case RichTextTags.ListItem:
+            <partial name="RichText/_ListItem" model="@Model" />
+            break;
+        case RichTextTags.Hyperlink:
+            <partial name="RichText/_Hyperlink" model="@Model" />
+            break;
+        case RichTextTags.Table:
+            <partial name="RichText/_Table" model="@Model" />
+            break;
+        case RichTextTags.TableRow:
+            <partial name="RichText/_TableRow" model="@Model" />
+            break;
+        case RichTextTags.TableHeaderCell:
+            <partial name="RichText/_TableHeaderCell" model="@Model" />
+            break;
+        case RichTextTags.TableCell:
+            <partial name="RichText/_TableCell" model="@Model" />
+            break;
+        case RichTextTags.Text:
+            <partial name="RichText/_Text" model="@Model" />
+            break;
+        default:
+            break;
+    }
+}
