+++ src/Dfe.ContentSupport.Web/Views/Shared/RichText/_TableCell.cshtml
@@ -1,0 +1,12 @@
+@using Newtonsoft.Json.Linq
+@model JToken
+
+<td>
+    @if (Model["content"] is JArray childNodes)
+    {
+        foreach (var childNode in childNodes)
+        {
+            <partial name="RichText/_RichText" model="@childNode" />
+        }
+    }
+</td>
