+++ src/Dfe.ContentSupport.Web/Views/Shared/RichText/_Text.cshtml
@@ -1,0 +1,22 @@
+@using Newtonsoft.Json.Linq
+@model JObject
+
+@{
+    var value = Model["value"]?.ToString();
+    var marks = Model["marks"] as JArray;
+
+    var isBold = marks?.Any(mark => mark["type"]?.ToString() == "bold") ?? false;
+
+    if (string.IsNullOrWhiteSpace(value))
+    {
+        @: <br />
+    }
+    else if (isBold)
+    {
+        <strong>@value</strong>
+    }
+    else
+    {
+        @value
+    }
+}
